Change parent to root, update tests.
The "parent" concept was in places implemented as an immediate parent,
but generally documented as referring to the complete representation.
To avoid this confusion, rename it "root" which unambiguously evokes
the full DataRep and not any intermediate level.

Rearrange the constructor parameters so that it is impossible to
supply conflicting values among the root and its fragments.
Add checks and exceptions for invalid parameter combinations.