Fix valid/unset checks, other review feedback.
Also consistently use "if self.fragment:" rather than "if self.root:"
for conditionals on being a fragment.  There is still one check on
self.root where the focus really is on the presence or absence of a root
as opposed to being a fragment (whether that involves self.root or not).

Rearranged the data property checks per code review feedback.

Update delete() for current view on handling fragments.