Update requirements, make work with rtd.
Our Read The Docs server likes gitlab.lab, not git.lab.
At least I think this is why the doc generation there won't work right now.
Also, we were using a github jsonpointer set up in September, but there
was a release to PyPI in October and nothing but a string formatting unit test
fix since then.  So I think we're good to switch to the regular pacakge.

In any event, the unit tests pass in a venv built with this file.