Add unit tests for bool.  Fix exposed bugs.
So unit tests are good.  They unearthed a relatively subtle bug in which
the first fetch of a deleted resource did *not* result in an exception,
but the subsequent ones do.  All of them should raise an exception.

Also restructured the nonzero implementation in terms of exceptions
rather than checking against validity, as UNSET does not behave consistently
between the two- you should always resolve an UNSET to determine truth,
while validity should leave it UNSET.