Stopgap HTTP exception solution.
This is a stopgap solution for getting distinct exceptions raised
for different HTTP statuses.  The details will need refinement
(to put it nicely).  The immediate goal is to allow users to write
code relying on the right set of exceptions instead of requests.HTTPError.

In particular, the functionality of making a DataRep out of it or otherwise
picking apart the requests.Response object is not being done now due to
lack of time.