Prevent **kwarg conflicts on follow/execute
Prevoiusly, follow could not have a kwarg 'name', and execute
could not have kwargs 'name' or 'data'.  This uses a single underscore
to avoid that (it does still prevent kwargs with that single underscore
prefix, but that it probably acceptable and matches the current
behavior of bind).

Note that this means that the data for an execute() must be passed
as a positional arg.  Hopefully people will not pass it as

BAD:
dr.execute('foo', _data={'a': 1})

and instead will do:

GOOD:
dr.execute('foo', {'a': 1})