Correctly dereference $ref schemas
We were trying to use the reference schema directly when indexing
through a reference, instead of following through to the referenced
schema.  This should really happen in reschema, and will be pushed
down in the near future.  This is a quick fix to unblock a client team.

Similarly, Multi schemas are problematic, but as in reschema we
defer them for now.  They weren't working anyway, now they fail to
work in a more obvious manner.